@jkowalski 2024-04-05
The gains are from pooling and recycling various decoders from https://github.com/klauspost/compress library using `Reset()`. This is saving memory allocations, sometimes goroutines and other initialization steps.